clinical-triage-env/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-triage-env"
version = "1.0.0"
description = "An OpenEnv environment simulating medical lab report triage in Indian diagnostic centers"
authors = [
    { name = "Nermal007" }
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "pydantic>=2.7.1",
    "openai>=1.30.1",
    "numpy>=1.26.4",
    "scipy>=1.13.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"
